fix: [model] was incorrectly using ``_report_type`` value that could be changed by inheriting classes.
For some reasons it seems that ``self._report_type`` was holding ``graphane``
value when I installed the ``report_graphane`` module. This could be due to a
change in inheriting rules.

However, we don't really need all this to find the correct report, so I removed
the use of ``self._report_type`` which seemed more as a safeguard against this
very bad way of registering reports and looking for them.